fix - bullet gravity bug
